requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
Jinja2==3.1.6
requests==2.32.3
python-dotenv==1.0.1

pymongo==4.11.0
certifi==2025.4.26

supabase==2.6.0

line-bot-sdk==3.17.1

gradio_client==1.9.0

pillow==11.2.1
numpy==2.2.6

python-multipart==0.0.20

PyJWT==2.10.1
passlib[bcrypt]==1.7.4

pydantic[email]==2.8.0